+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 
+1.1.3 / 2018-11-16
+==================
+
+  * fixing deprecation on error route from @mansona
+
+1.1.2 / 2018-11-16
+==================
+
+  * adding tests and removed deprecation for ember-routing.route-router #26 from @mansona
+
+1.1.1 / 2018-11-16
+==================
+
+  * fixing loading class on body from @mansona
+
 1.1.0 / 2018-10-30
 ==================